♻️ Refactor Menu, split menus into plugins, and have one load 'em all
