+++ src/menu/main_menu.rs
@@ -1,38 +1,29 @@
-use crate::{GameState, sound::AudioSettings, ui::sliders::base_slider};
-use bevy::{
-    prelude::*,
-    ui_widgets::{SliderValue, ValueChange, observe},
-};
+use crate::GameState;
+use bevy::prelude::*;
 
 #[derive(Component)]
-pub struct PlayButton;
+struct PlayButton;
 
 #[derive(Component)]
-pub struct QuitButton;
+struct QuitButton;
 
 #[derive(Component)]
-pub struct CreditsButton;
+struct CreditsButton;
 
 #[derive(Component)]
-pub struct TutorialButton;
+struct TutorialButton;
 
 #[derive(Component)]
-pub struct BackButton;
+struct BackButton;
 
 #[derive(Component)]
-pub struct SettingsButton;
+struct SettingsButton;
 
-#[derive(Component)]
-pub struct SettingsUI;
+pub struct MainMenuPlugin;
 
-#[derive(Component)]
-pub struct VolumeSlider;
-
-#[derive(Component)]
-pub struct VolumeText;
-
-pub fn main_menu_plugin(app: &mut App) {
-    app.add_systems(OnEnter(GameState::MainMenu), setup_main_menu)
+impl Plugin for MainMenuPlugin {
+    fn build(&self, app: &mut App) {
+        app.add_systems(OnEnter(GameState::MainMenu), setup_main_menu)
         .add_systems(
             Update,
             (
@@ -45,18 +36,12 @@ pub fn main_menu_plugin(app: &mut App) {
             )
                 .run_if(in_state(GameState::MainMenu)),
         )
-        .add_systems(OnEnter(GameState::Credits), setup_credits)
-        .add_systems(
+         .add_systems(
             Update,
             (button_system, handle_back_button)
                 .run_if(in_state(GameState::Credits).or(in_state(GameState::Settings))),
-        )
-        .add_systems(OnEnter(GameState::Settings), setup_settings)
-        .add_systems(
-            Update,
-            sync_volume_slider.run_if(in_state(GameState::Settings)),
-        )
-        .add_systems(Update, update_volume_text);
+        );
+    }
 }
 
 fn setup_main_menu(mut commands: Commands) {
@@ -214,75 +199,6 @@ fn setup_main_menu(mut commands: Commands) {
         });
 }
 
-fn setup_credits(mut commands: Commands) {
-    commands.spawn((Camera2d, DespawnOnExit(GameState::Credits)));
-    commands
-        .spawn((
-            DespawnOnExit(GameState::Credits),
-            Node {
-                width: Val::Percent(100.0),
-                height: Val::Percent(100.0),
-                align_items: AlignItems::Center,
-                justify_content: JustifyContent::Center,
-                flex_direction: FlexDirection::Column,
-                ..default()
-            },
-            BackgroundColor(Color::BLACK),
-        ))
-        .with_children(|parent| {
-            parent.spawn((
-                Text::new(
-                    r#"
-Symulator Reaktora z Urankiem
-
-Autorzy:
-
-Kacper Sowinski - Project Manager, Developer
-Alan Klas - Lead Developer, Code Reviewer, 
-Mateusz Oskar Kmiec - Developer, Sound & Visual Designer,
-Ignacy Sztykiel - Developer, Sound & Visual Designer
-                
-                "#,
-                ),
-                TextFont {
-                    font_size: 32.0,
-                    ..default()
-                },
-                TextColor(Color::WHITE),
-                Node {
-                    margin: UiRect::all(Val::Px(20.0)),
-                    justify_content: JustifyContent::Center,
-                    align_items: AlignItems::Center,
-                    ..default()
-                },
-            ));
-            parent
-                .spawn((
-                    Button,
-                    Node {
-                        width: Val::Px(150.0),
-                        height: Val::Px(50.0),
-                        justify_content: JustifyContent::Center,
-                        align_items: AlignItems::Center,
-                        margin: UiRect::all(Val::Px(20.0)),
-                        ..default()
-                    },
-                    BackgroundColor(Color::srgb(0.2, 0.2, 0.2)),
-                    BackButton,
-                ))
-                .with_children(|p| {
-                    p.spawn((
-                        Text::new("Back"),
-                        TextFont {
-                            font_size: 24.0,
-                            ..default()
-                        },
-                        TextColor(Color::WHITE),
-                    ));
-                });
-        });
-}
-
 #[allow(clippy::type_complexity)]
 fn button_system(
     mut interaction_query: Query<
@@ -400,135 +316,4 @@ fn handle_settings_button(
             next_state.set(GameState::Settings);
         }
     }
-}
-
-pub fn update_volume_text(
-    settings: Res<AudioSettings>,
-    mut texts: Query<&mut Text, With<VolumeText>>,
-) {
-    if !settings.is_changed() {
-        return;
-    }
-
-    for mut text in texts.iter_mut() {
-        **text = format!("{}%", (settings.volume * 100.0) as i32);
-    }
-}
-
-pub fn sync_volume_slider(
-    settings: Res<AudioSettings>,
-    sliders: Query<(Entity, &SliderValue), With<VolumeSlider>>,
-    mut commands: Commands,
-) {
-    if settings.is_changed() {
-        for (entity, value) in sliders.iter() {
-            if value.0 != settings.volume {
-                commands.entity(entity).insert(SliderValue(settings.volume));
-            }
-        }
-    }
-}
-
-pub fn create_volume_slider(initial_value: f32) -> impl Bundle {
-    (
-        VolumeSlider,
-        base_slider(initial_value, 1.0),
-        observe(
-            |value_change: On<ValueChange<f32>>, mut settings: ResMut<AudioSettings>| {
-                settings.volume = value_change.value;
-            },
-        ),
-    )
-}
-
-fn setup_settings(
-    mut commands: Commands,
-    asset_server: Res<AssetServer>,
-    settings: Res<AudioSettings>,
-) {
-    commands.spawn((Camera2d, DespawnOnExit(GameState::Settings)));
-    commands
-        .spawn((
-            DespawnOnExit(GameState::Settings),
-            Node {
-                width: Val::Percent(100.0),
-                height: Val::Percent(100.0),
-                flex_direction: FlexDirection::Column,
-                padding: UiRect::all(Val::Px(32.0)),
-                ..default()
-            },
-            BackgroundColor(Color::srgba(0.0, 0.0, 0.0, 0.75)),
-            SettingsUI,
-        ))
-        .with_children(|parent| {
-            parent.spawn((
-                Text::new("Settings"),
-                TextFont {
-                    font: asset_server.load("fonts/LTSuperior-Medium.ttf"),
-                    font_size: 32.0,
-                    ..default()
-                },
-            ));
-
-            parent
-                .spawn((Node {
-                    width: Val::Percent(50.0),
-                    margin: UiRect {
-                        top: Val::Px(32.0),
-                        ..default()
-                    },
-                    row_gap: Val::Px(12.0),
-                    flex_direction: FlexDirection::Column,
-                    ..default()
-                },))
-                .with_children(|parent| {
-                    parent.spawn((
-                        Text::new("Volume"),
-                        TextFont {
-                            font: asset_server.load("fonts/LTSuperior-Regular.ttf"),
-                            font_size: 24.0,
-                            ..default()
-                        },
-                    ));
-
-                    parent
-                        .spawn(Node {
-                            flex_direction: FlexDirection::Row,
-                            column_gap: Val::Px(12.0),
-                            ..default()
-                        })
-                        .with_children(|parent| {
-                            parent.spawn(create_volume_slider(settings.volume));
-                            parent.spawn((
-                                Text::new(format!("{}%", (settings.volume * 100.0) as u8)),
-                                VolumeText,
-                            ));
-                        });
-
-                    parent
-                        .spawn((
-                            Button,
-                            Node {
-                                width: Val::Px(150.0),
-                                height: Val::Px(50.0),
-                                justify_content: JustifyContent::Center,
-                                align_items: AlignItems::Center,
-                                margin: UiRect::all(Val::Px(20.0)),
-                                ..default()
-                            },
-                            BackgroundColor(Color::srgb(0.2, 0.2, 0.2)),
-                            BackButton,
-                        ))
-                        .with_children(|parent| {
-                            parent.spawn((
-                                Text::new("Back"),
-                                TextFont {
-                                    font_size: 24.0,
-                                    ..default()
-                                },
-                                TextColor(Color::WHITE),
-                            ));
-                        });
-                });
-        });
 }
