+++ src/menu/menu.rs
@@ -1,0 +1,14 @@
+use bevy::prelude::*;
+
+use crate::menu::{credits::CreditsMenuPlugin, main_menu::MainMenuPlugin, settings::SettingsMenuPlugin};
+
+pub struct MenuPlugin;
+
+impl Plugin for MenuPlugin {
+    fn build(&self, app: &mut App) {
+        app
+        .add_plugins(CreditsMenuPlugin)
+        .add_plugins(SettingsMenuPlugin)
+        .add_plugins(MainMenuPlugin);
+    }
+}
