+++ src/menu/mod.rs
@@ -1,0 +1,5 @@
+mod credits;
+mod settings;
+mod main_menu;
+mod menu;
+pub use menu::MenuPlugin;
